+++ src/flow/src/error.rs
@@ -21,7 +21,6 @@ use common_error::ext::BoxedError;
 use common_macro::stack_trace_debug;
 use common_telemetry::common_error::ext::ErrorExt;
 use common_telemetry::common_error::status_code::StatusCode;
-use datatypes::prelude::ConcreteDataType;
 use snafu::{Location, Snafu};
 
 use crate::adapter::FlowId;
@@ -42,18 +41,6 @@ pub enum Error {
     #[snafu(display("Internal error"))]
     Internal {
         reason: String,
-        #[snafu(implicit)]
-        location: Location,
-    },
-
-    #[snafu(display(
-        "Flow output type mismatch: expected: {:?}, actual: {:?}",
-        expected,
-        actual
-    ))]
-    FlowOutputTypeMismatch {
-        expected: ConcreteDataType,
-        actual: ConcreteDataType,
         #[snafu(implicit)]
         location: Location,
     },
@@ -226,9 +213,7 @@ impl ErrorExt for Error {
                 source.status_code()
             }
             Self::MetaClientInit { source, .. } => source.status_code(),
-            Self::ParseAddr { .. } | Self::FlowOutputTypeMismatch { .. } => {
-                StatusCode::InvalidArguments
-            }
+            Self::ParseAddr { .. } => StatusCode::InvalidArguments,
         }
     }
