+++ tests/cases/standalone/common/flow/flow_basic.result
@@ -1047,15 +1047,15 @@ FROM
 
 -- Test if FLOWS table works, but don't care about the result since it vary from runs
 SELECT
-    1
+    count(CASE WHEN state_size > 0 THEN 1 ELSE 0 END) as active_flows,
 FROM
     INFORMATION_SCHEMA.FLOWS;
 
-+----------+
-| Int64(1) |
-+----------+
-| 1        |
-+----------+
++--------------+
+| active_flows |
++--------------+
+| 1            |
++--------------+
 
 DROP FLOW requests_long_term;
